% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_BCCh_data.R
\name{get_bcch_data}
\alias{get_bcch_data}
\title{Fetch data using an .iqy file from the BCCh}
\usage{
get_bcch_data(path_to_iqy, q_values = NULL)
}
\arguments{
\item{path_to_iqy}{the path to the .iqy file. Can be relative to wd.}

\item{q_values}{a character vector (or something that can be coerced to it).
Should contain the values for every parameter in the query, in adequate
order. Use only if you already know the query and its parameters. If left
\code{NULL}, the program will ask the user to input these values.}
}
\value{
a data frame with attribute \code{data_types}, which is a character
    vector that shows the kind of data stored in each column.
}
\description{
\code{get_bcch_data} parses an .iqy file generated by the Statistical Database
    of the Central Bank of Chile (BCCh), gets the required data from their server,
    and returns it as a data frame.
}
\examples{

\dontrun{
# this should ask the use for start and end date
get_bcch_data("UF_IVP_DIARIO.iqy")
}

}
\seealso{
Vist the BCCh database at \url{https://si3.bcentral.cl/Siete/secure/cuadros/home.aspx}.
}
