% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_BCCh_data.R
\name{get_bcch_data}
\alias{get_bcch_data}
\title{Fetch data using an .iqy file from the BCCh}
\usage{
get_bcch_data(path_to_iqy, q_values, ...)
}
\arguments{
\item{path_to_iqy}{the path to the .iqy file. Can be a relative path.}

\item{q_values}{a character vector (or something that can be coerced to it).
Should contain the values for every parameter in the query, in adequate
order. Parameters are almost always start and end date, but use only if
you are sure of this. If missing, the program will ask the user
to input these values.}

\item{...}{arguments passed to \code{httr::POST}. For example, you can pass
\code{httr::timeout(20)} to wait 20 seconds for the response if the default
is not enough.}
}
\value{
A data frame containing the requested data and with the following
    attributes:
    \enumerate{
        \item{\code{data_def}: definition of the data requested}
        \item{\code{data_types}: type of data stored in each column}
    }
}
\description{
\code{get_bcch_data} parses an .iqy file generated by the Statistical Database
    of the Central Bank of Chile (BCCh), gets the required data from their server,
    and returns it as a data frame.
}
\examples{

\dontrun{
# this should ask the user for a start and an end date
get_bcch_data("UF_IVP_DIARIO.iqy")
}

}
\seealso{
Vist the BCCh database at \url{https://si3.bcentral.cl/siete}.
}
